feat: use fake xhr with `data/users.json`
